+++ kernel/cmd/time.c
@@ -11,7 +11,7 @@
 #include <lib/str.h>
 #include <lib/num.h>
 
-void set_time(uint8_t hours, uint8_t minutes) {
+void set_time(uint32_t hours, uint32_t minutes) {
     if (hours >= 24 || minutes >= 60) {
         printf("Invalid time format! Hours must be 0-23, minutes must be 0-59\n");
         return;
@@ -25,8 +25,8 @@ void set_time(uint8_t hours, uint8_t minutes) {
 
 void time_cmd(char** args) {
     if (args[1] && args[2]) {
-        uint8_t hours = str_to_uint8(args[1]);
-        uint8_t minutes = str_to_uint8(args[2]);
+        uint32_t hours = str_to_uint8(args[1]);
+        uint32_t minutes = str_to_uint8(args[2]);
         
         if (hours >= 24 || minutes >= 60) {
             printf("Invalid time format! Hours must be 0-23, minutes must be 0-59\n");
